Refactor: 칸반보드 기능 코드 리뷰에 맞춘 코드 리팩토링

+++ src/main/java/com/codiapp/codi/domain/project/service/project/ProjectService.java
@@ -12,6 +12,6 @@ public interface ProjectService {
 
 	  void updateProject(Project project);
 	  
-	  void deleteProject(int id);
+	  void deleteProject(Long id);
 
 }
